Minors display correctings, bugfix.
